fix: problems in stc router
