+++ src/cmd/index.ts
@@ -2,30 +2,6 @@ import commander from "commander";
 import { CommandPlugin } from "./base";
 import { RouterPlugin } from "./router";
 import { DevPlugin } from "./dev";
-import {
-  CmdConfig,
-  createCmdConfig,
-  mergeCmdConfig,
-  RouterConfig as RConfig,
-  IENV as E
-} from "./base";
-
-export namespace CMD {
-  export interface Env extends E {}
-  export interface RouterConfig extends RConfig {}
-  export interface Config extends CmdConfig {
-    env?: Env;
-    routers?: RouterConfig;
-  }
-
-  export function create(config: Config) {
-    return createCmdConfig(config);
-  }
-
-  export function merge(merge: Config, config: Config) {
-    return mergeCmdConfig(merge, config);
-  }
-}
 
 function initCommand(plugin: CommandPlugin) {
   const program = commander.name(plugin.name).description(plugin.description);
